Leiningen: remove portal require as library not supported
Remove portal.api require from main notespace.  Add a require for portal in the
rich comment block at the end of the code.

Running a REPL with Leiningen causes issues with portal library, so removing it
from namespaces that Leiningen reads on startup.  This stops issues when
starting the repl and evaluating portal code.

Clojure CLI uses the `dev/user.clj` file to require portal, a file that
Leiningen is not configured to read.